Rewrite create and altertable functions to be hook based
