+++ src/alphadb/src/core/query/mod.rs
@@ -2,3 +2,4 @@ pub mod build;
 pub mod column;
 pub mod default_data;
 pub mod primary_key;
+pub mod table;
