+++ src/alphadb/src/engine/postgres_impl/query/table/mod.rs
@@ -13,5 +13,4 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-pub mod altertable;
-pub mod createtable;
+pub mod config;
